refactor: deleted header.component, deleted inline styles
