+++ src/app/app.routes.ts
@@ -1,6 +1,5 @@
 import { Routes } from '@angular/router';
 import { UsersListComponent } from './users-list/users-list.component';
-import { HeaderComponent } from './header/header.component';
 import { ContentComponent } from './content/content.component';
 import { TodoListComponent } from './todo-list/todo-list.component';
