+++ src/app/content/content.component.html
@@ -10,7 +10,7 @@
         <img src="../assets/Bricks.svg" alt="" />
       </div>
       <div class="description">
-        <b style="color: #3c3c3b">Большой ассортимент</b> строительных материалов
+        <b >Большой ассортимент</b> строительных материалов
         всегда в наличии
       </div>
     </div>
@@ -19,7 +19,7 @@
         <img src="../assets/Car.svg" alt="" />
       </div>
       <div class="description">
-        <b style="color: #3c3c3b">Доставка стройматериалов</b>
+        <b>Доставка стройматериалов</b>
         в течении четырех часов или в любое удобное время
       </div>
     </div>
@@ -29,7 +29,7 @@
       </div>
       <div class="description">
         Многоуровневая
-        <b style="color: #3c3c3b">система управления качеством</b> по стандарту
+        <b>система управления качеством</b> по стандарту
         ISO 9001
       </div>
     </div>
@@ -97,7 +97,6 @@
           <img
             src="../assets/heart_card.svg"
             alt="heart"
-            style="cursor: pointer"
           />
         </div>
         <div class="good_pictire">
@@ -114,7 +113,7 @@
           <div class="amount">
             <div class="amount_text">Количество</div>
             <span>-</span>
-            <span style="font-size: 20px">1</span>
+            <span>1</span>
             <span>+</span>
           </div>
           <div class="basket_button">
@@ -128,7 +127,6 @@
           <img
             src="../assets/heart_card.svg"
             alt="heart"
-            style="cursor: pointer"
           />
         </div>
         <div class="good_pictire">
@@ -145,7 +143,7 @@
           <div class="amount">
             <div class="amount_text">Количество</div>
             <span>-</span>
-            <span style="font-size: 20px">1</span>
+            <span>1</span>
             <span>+</span>
           </div>
           <div class="basket_button">
@@ -160,7 +158,6 @@
           <img
             src="../assets/heart_card.svg"
             alt="heart"
-            style="cursor: pointer"
           />
         </div>
         <div class="good_pictire">
@@ -177,7 +174,7 @@
           <div class="amount">
             <div class="amount_text">Количество</div>
             <span>-</span>
-            <span style="font-size: 20px">1</span>
+            <span>1</span>
             <span>+</span>
           </div>
           <div class="basket_button">
@@ -191,7 +188,6 @@
           <img
             src="../assets/heart_card.svg"
             alt="heart"
-            style="cursor: pointer"
           />
         </div>
         <div class="good_pictire">
@@ -207,7 +203,7 @@
           <div class="amount">
             <div class="amount_text">Количество</div>
             <span>-</span>
-            <span style="font-size: 20px">1</span>
+            <span>1</span>
             <span>+</span>
           </div>
           <div class="basket_button">
@@ -239,7 +235,6 @@
           <img
             src="../assets/heart_card.svg"
             alt="heart"
-            style="cursor: pointer"
           />
         </div>
         <div class="good_pictire">
@@ -255,7 +250,7 @@
           <div class="amount">
             <div class="amount_text">Количество</div>
             <span>-</span>
-            <span style="font-size: 20px">1</span>
+            <span>1</span>
             <span>+</span>
           </div>
           <div class="basket_button">
@@ -269,7 +264,6 @@
           <img
             src="../assets/heart_card.svg"
             alt="heart"
-            style="cursor: pointer"
           />
         </div>
         <div class="good_pictire">
@@ -285,7 +279,7 @@
           <div class="amount">
             <div class="amount_text">Количество</div>
             <span>-</span>
-            <span style="font-size: 20px">1</span>
+            <span>1</span>
             <span>+</span>
           </div>
           <div class="basket_button">
@@ -299,7 +293,6 @@
           <img
             src="../assets/heart_card.svg"
             alt="heart"
-            style="cursor: pointer"
           />
         </div>
         <div class="good_pictire">
@@ -315,7 +308,7 @@
           <div class="amount">
             <div class="amount_text">Количество</div>
             <span>-</span>
-            <span style="font-size: 20px">1</span>
+            <span>1</span>
             <span>+</span>
           </div>
           <div class="basket_button">
@@ -329,7 +322,6 @@
           <img
             src="../assets/heart_card.svg"
             alt="heart"
-            style="cursor: pointer"
           />
         </div>
         <div class="good_pictire">
@@ -345,7 +337,7 @@
           <div class="amount">
             <div class="amount_text">Количество</div>
             <span>-</span>
-            <span style="font-size: 20px">1</span>
+            <span>1</span>
             <span>+</span>
           </div>
           <div class="basket_button">
