--- src/app/header/header.component.html
@@ -1,59 +1,0 @@
-<div class="header">
-  <div class="header_cont">
-    <div class="header_left">
-      <div class="logo">
-        <img src="../assets/Logo.svg" alt="logo" />
-      </div>
-      <nav>
-        <ul>
-          <li style="color: orange" *ngIf="isShowMain" routerLink="/">
-            {{ header_item_1 }}
-          </li>
-          <li>{{ aboutCompany }}</li>
-          <li>{{ header_item_3 }}</li>
-          <li style="color: orange" routerLink="users">Пользователи</li>
-          <li style="color: orange" routerLink="todos">Список дел</li>
-          <button (click)="isShowMain = !isShowMain">Скрыть/Показать</button>
-        </ul>
-      </nav>
-    </div>
-    <div class="header_right">
-      <div class="contacts">
-        <div class="phone_logo">
-          <img src="../assets/Phone.svg" alt="phone_logo" />
-        </div>
-        <div class="number">+7 (965) 084-29-29</div>
-        <div class="heart_logo">
-          <img src="../assets/heart.svg" alt="heart_logo" />
-        </div>
-      </div>
-      <div class="basket">
-        <button>
-          <img class="basket_logo" src="../assets/cart.svg" alt="basket" />
-          <span>Пусто</span>
-        </button>
-      </div>
-    </div>
-  </div>
-</div>
-<div class="header_second">
-  <div class="cont_second">
-    <nav>
-      <img src="../assets/light.svg" alt="catalogue" />
-      <ul>
-        <li>
-          {{
-            isUpperCase ? lowerCaseItems.join(" ") : upperCaseItems.join(" ")
-          }}
-        </li>
-      </ul>
-      <button (click)="isUpperCase = !isUpperCase">Изменить регистр</button>
-    </nav>
-    <div class="search">
-      <input placeholder="Введите для поиска ..." />
-      <img src="../assets/Mask group.svg" alt="lupa" />
-      <img class="location" src="../assets/map-pin (2) 1.svg" alt="" />
-      <span class="city">Москва</span>
-    </div>
-  </div>
-</div>
